visual-e-reader/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
opencv-python==4.8.1.78
numpy>=1.26.0
python-dotenv==1.0.0
httpx==0.25.2
pydantic==2.5.0
replicate
